More removals for 2.0
