+++ lib/graphql/types/relay.rb
@@ -14,8 +14,6 @@ require "graphql/types/relay/page_info"
 require "graphql/types/relay/base_connection"
 require "graphql/types/relay/base_edge"
 require "graphql/types/relay/node"
-require "graphql/types/relay/node_field"
-require "graphql/types/relay/nodes_field"
 
 module GraphQL
   module Types
